Fix hot reloading.
