--- src/App.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-
-import BlogPage from './components/containers/BlogPage';
-
-const App = () => (
-  <MuiThemeProvider>
-    <BlogPage />
-  </MuiThemeProvider>
-);
-
-export default App;
